+++ transports/smallwebrtc/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to **Pipecat Client React Native Small WebRTC** will be docu
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4.1]
+
+- Adding support for the `onLocalAudioLevel` callback.
+- Adding support for the `onRemoteAudioLevel` callback.
+
+## [1.4.0] - 2025-11-07
 
 - `SmallWebRTCTransport` for React Native, matching all the features provided by Pipecat Client 1.4.0.